feat: add multiple ravens in canvas animation
This commit introduces the ability to generate and animate multiple ravens in the canvas instead of a single raven. This is achieved by using a Raven class and an array to manage multiple instances of the Raven class. This feature enhances the visual appeal and complexity of the canvas animation.