Add random trail to Raven class and update Particle radius
This commit introduces a new feature where the Raven class now has a random chance to have a trailing effect. The radius of the particles in the Particle class has also been increased for a more pronounced effect. Rendering order has also been updated to prioritize particles, ravens, and explosions in that order.